. modules are now aware of the sampling rate . doubles can be now used in the mdl's data structures . fixes to ondemand querying . adding code to profile capture, disabled by default

+++ src/branches/2.0/modules/autofocus/data.h
@@ -34,7 +34,7 @@
 
 como_tuple struct autofocus_tuple {
     uint32_t addr;
-    uint64_t bytes;
+    double bytes;
 };
 
 como_record struct autofocus_record {
